requirements.txt
streamlit
tensorflow
numpy
plost
pillow
